rename Nav to NavBar to follow Bulma naming
