+++ src-ui/lib/components/core/NavBar.svelte
